move 'set_up_configuration' function to tools/options.py
